Vim working with a C and C-c solution that's fast and comfy.
Make sure in MacOSX to set Caps Lock to no action, otherwise it will
make a weird silent toggling caps lock. The normal C-[ works to exit,
but its slow because of waiting for the next command. So C-c is faster,
nice to have both. Also added shortcuts to try for save and save-quit.